Issue #1586: moved disable desktop mode on host change to SessionRepo
